Author Info Not a Search Result
Many search results returned Error: 'NoneType' object has no attribute
'text’.  This was because the way the library was currently coded after
the last link the traversal would continue into the custom area on the
right displaying author, actor and other such information.  This area
did not have a description and the blank description would error out
the library.  Results without descriptions are now not included in the
search results.